Refactor HammerTechnologyTest into a new file
Close #260
